added more file exist checks
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@10498 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ jenkins-dist.sh
@@ -238,9 +238,17 @@ do
   fi
 done
 
-cp $DISTPATH/phreeqc3-doc/RELEASE.TXT          $DISTPATH/doc/RELEASE
-cp $DISTPATH/phreeqc3-doc/NOTICE.TXT           $DISTPATH/doc/NOTICE
-cp $DISTPATH/phreeqc3-doc/README.IPhreeqc.TXT  $DISTPATH/doc/README
+if [ -e $DISTPATH/phreeqc3-doc/RELEASE.TXT ]; then
+  cp $DISTPATH/phreeqc3-doc/RELEASE.TXT         $DISTPATH/doc/RELEASE
+fi
+
+if [ -e $DISTPATH/phreeqc3-doc/NOTICE.TXT ]; then
+  cp $DISTPATH/phreeqc3-doc/NOTICE.TXT          $DISTPATH/doc/NOTICE
+fi
+
+if [ -e $DISTPATH/phreeqc3-doc/README.IPhreeqc.TXT ]; then
+  cp $DISTPATH/phreeqc3-doc/README.IPhreeqc.TXT $DISTPATH/doc/README
+fi
 
 ##(cd "$DISTPATH/doc" && "doxygen")
 ##
